Address review feedback
* Extract GrpcAPIService to its own file and clarify comments on why
  do not generate it from the proto file
* Consistently access Registry.externalHTTPRules through accessors.
  For this the AddExternalHTTPRule setter was added
* Renamed registerGrpcAPIService to registerHTTPRulesFromGrpcAPIService
* Replace use of optss as plural of opts with optsList. Cannot use
  suggested options because of name collision.
