+++ protoc-gen-grpc-gateway/descriptor/grpc_api_configuration.go
@@ -8,30 +8,7 @@ import (
 
 	"github.com/ghodss/yaml"
 	"github.com/golang/protobuf/jsonpb"
-	"github.com/golang/protobuf/proto"
-	"google.golang.org/genproto/googleapis/api/annotations"
 )
-
-// GrpcAPIService represents a stripped down version of google.api.Service .
-// Compare to https://github.com/googleapis/googleapis/blob/master/google/api/service.proto
-//
-// Note that for the purposes of the gateway generator we only consider a subset of all
-// available features google supports in their service descriptions and hence we do not
-// bother having the full service protobuf which has a lot of other dependencies. Thanks
-// to backwards compatibility this should be relatively safe.
-type GrpcAPIService struct {
-	// Http Rule. Named Http in the actual proto. Changed to suppress linter warning.
-	HTTP *annotations.Http `protobuf:"bytes,9,opt,name=http" json:"http,omitempty"`
-}
-
-// ProtoMessage returns an empty GrpcAPIService element
-func (*GrpcAPIService) ProtoMessage() {}
-
-// Reset resets the GrpcAPIService
-func (m *GrpcAPIService) Reset() { *m = GrpcAPIService{} }
-
-// String returns the string representation of the GrpcAPIService
-func (m *GrpcAPIService) String() string { return proto.CompactTextString(m) }
 
 func loadGrpcAPIServiceFromYAML(yamlFileContents []byte, yamlSourceLogName string) (*GrpcAPIService, error) {
 	jsonContents, err := yaml.YAMLToJSON(yamlFileContents)
@@ -52,7 +29,7 @@ func loadGrpcAPIServiceFromYAML(yamlFileContents []byte, yamlSourceLogName strin
 	return &serviceConfiguration, nil
 }
 
-func registerGrpcAPIService(registry *Registry, service *GrpcAPIService, sourceLogName string) error {
+func registerHTTPRulesFromGrpcAPIService(registry *Registry, service *GrpcAPIService, sourceLogName string) error {
 	if service.HTTP == nil {
 		// Nothing to do
 		return nil
@@ -64,14 +41,14 @@ func registerGrpcAPIService(registry *Registry, service *GrpcAPIService, sourceL
 			return fmt.Errorf("Selector '%v' in %v must specify a single service method without wildcards", rule.GetSelector(), sourceLogName)
 		}
 
-		registry.externalHTTPRules[selector] = append(registry.externalHTTPRules[selector], rule)
+		registry.AddExternalHTTPRule(selector, rule)
 	}
 
 	return nil
 }
 
 // LoadGrpcAPIServiceFromYAML loads a gRPC API Configuration from the given YAML file
-// and registers the HttpRule descriptions contained in it  as externalHTTPRules in
+// and registers the HttpRule descriptions contained in it as externalHTTPRules in
 // the given registry. This must be done before loading the proto file.
 //
 // You can learn more about gRPC API Service descriptions from google's documentation
@@ -90,5 +67,5 @@ func (r *Registry) LoadGrpcAPIServiceFromYAML(yamlFile string) error {
 		return err
 	}
 
-	return registerGrpcAPIService(r, service, yamlFile)
+	return registerHTTPRulesFromGrpcAPIService(r, service, yamlFile)
 }
